+++ Documentation/API/FingerController.InputType.md.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9bcbf770dda5e3e48bca49d8dcd1c776
+guid: 7929be1ab61cb0746b7beedce65f9ff9
 TextScriptImporter:
   externalObjects: {}
   userData:
