+++ Documentation/API/FingerController.md.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5148dbfa57335974b90fa809c78dbf01
+guid: c8ac8fbe82f42a84cb43dddde7dd6dd1
 TextScriptImporter:
   externalObjects: {}
   userData:
