+++ Documentation/API/HandFacade.md
@@ -1,0 +1,96 @@
+# Class HandFacade
+
+The public facade for the hand.
+
+## Contents
+
+* [Inheritance]
+* [Namespace]
+* [Syntax]
+* [Properties]
+  * [IndexFinger]
+  * [MiddleFinger]
+  * [PinkyFinger]
+  * [RingFinger]
+  * [Thumb]
+
+## Details
+
+##### Inheritance
+
+* System.Object
+* HandFacade
+
+##### Namespace
+
+* [Tilia.Visuals.BasicHand]
+
+##### Syntax
+
+```
+public class HandFacade : MonoBehaviour
+```
+
+### Properties
+
+#### IndexFinger
+
+The [FingerController] for the Index Finger.
+
+##### Declaration
+
+```
+public FingerController IndexFinger { get; set; }
+```
+
+#### MiddleFinger
+
+The [FingerController] for the Middle Finger.
+
+##### Declaration
+
+```
+public FingerController MiddleFinger { get; set; }
+```
+
+#### PinkyFinger
+
+The [FingerController] for the Pinky Finger.
+
+##### Declaration
+
+```
+public FingerController PinkyFinger { get; set; }
+```
+
+#### RingFinger
+
+The [FingerController] for the Ring Finger.
+
+##### Declaration
+
+```
+public FingerController RingFinger { get; set; }
+```
+
+#### Thumb
+
+The [FingerController] for the Thumb.
+
+##### Declaration
+
+```
+public FingerController Thumb { get; set; }
+```
+
+[Tilia.Visuals.BasicHand]: README.md
+[FingerController]: FingerController.md
+[Inheritance]: #Inheritance
+[Namespace]: #Namespace
+[Syntax]: #Syntax
+[Properties]: #Properties
+[IndexFinger]: #IndexFinger
+[MiddleFinger]: #MiddleFinger
+[PinkyFinger]: #PinkyFinger
+[RingFinger]: #RingFinger
+[Thumb]: #Thumb
