+++ Documentation/API/HandFacade.md.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f1def9424ef1c8048b6693d77c00f265
+guid: 3b7fa37c08ecd3345ae6c1369d6cbe9e
 TextScriptImporter:
   externalObjects: {}
   userData:
