+++ Runtime/SharedResources/Scripts/FingerController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 22e9e1d746f46ef47a64cee172b1e3f5
+guid: eb4be2867c44b424491317d2096fb18b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
