+++ Runtime/SharedResources/Scripts/HandFacade.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ce7a4f8120a098444b753a63bc32e69f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
